Moved E2E TestVSchemaTrackerInit test to misc
